M5-10 Refactor CLI commands into separate modules
Moved collection, count, find, stats, and test command logic from src/app.js into dedicated files under src/commands/. Updated app.js to use these new command handlers, improving code organization and maintainability.
